refactor and add UT and add scenario test
