+++ src/spring-cloud/azext_spring_cloud/app_managed_identity.py
@@ -1,0 +1,296 @@
+# --------------------------------------------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License. See License.txt in the project root for license information.
+# --------------------------------------------------------------------------------------------
+
+from ._utils import wait_till_end
+from .vendored_sdks.appplatform.v2022_03_01_preview import models as models_20220301preview
+from azure.cli.core.commands import arm as _arm
+from knack.log import get_logger
+from knack.util import CLIError
+
+
+logger = get_logger(__name__)
+
+
+def app_identity_assign(cmd,
+                        client,
+                        resource_group,
+                        service,
+                        name,
+                        role=None,
+                        scope=None,
+                        system_assigned=None,
+                        user_assigned=None):
+    """
+    Note: Always use sync method to operate managed identity to avoid data inconsistency.
+    :param role: role name of role assignment for system-assigned managed identity.
+    :param scope: scope of role assignment for system-assigned managed identity.
+    :param system_assigned: 1. None or False: Don't change system-assigned managed identity.
+                            2. Enable system-assigned managed identity on app.
+    :param user_assigned: 1. None: Don't change user-assigned managed identities.
+                          2. A non-empty list of user-assigned managed identity resource id to app.
+                          3. A empty list: should be blocked by validator.
+    """
+    # TODO(jiec): Retire legacy identity assign after migration.
+    if _is_legacy_identity_assign(system_assigned, user_assigned):
+        _legacy_app_identity_assign(cmd, client, resource_group, service, name)
+    else:
+        _new_app_identity_assign(cmd, client, resource_group, service, name, system_assigned, user_assigned)
+
+    if role and scope:
+        _create_role_assignment(resource_group, service, name, role, scope)
+
+    return client.apps.get(resource_group, service, name)
+
+
+def app_identity_remove(cmd,
+                        client,
+                        resource_group,
+                        service,
+                        name,
+                        system_assigned=None,
+                        user_assigned=None):
+    """
+    Note: Always use sync method to operate managed identity to avoid data inconsistency.
+    :param system_assigned: 1) None or False: Don't change system-assigned managed identity.
+                            2) True: remove system-assigned managed identity
+    :param user_assigned: 1) None: Don't change user-assigned managed identities.
+                          2) An empty list: remove all user-assigned managed identities.
+                          3) A non-empty list of user-assigned managed identity resource id to remove.
+    """
+    app = client.apps.get(resource_group, service, name)
+
+    if not app.identity:
+        logger.debug("Skip remove managed identity since no identities assigned to app.")
+        return
+    if not app.identity.type:
+        raise CLIError("Invalid existed identity type {}.".format(app.identity.type))
+    if app.identity.type == models_20220301preview.ManagedIdentityType.NONE:
+        logger.debug("Skip remove managed identity since identity type is {}.".format(app.identity.type))
+        return
+
+    # TODO(jiec): For back-compatible, convert to remove system-assigned only case. Remove code after migration.
+    if system_assigned is None and user_assigned is None:
+        system_assigned = True
+
+    new_user_identities = _get_new_user_identities_for_remove(app.identity.user_assigned_identities, user_assigned)
+    new_identity_type = _get_new_identity_type_for_remove(app.identity.type, system_assigned, new_user_identities)
+    user_identity_payload = _get_user_identity_payload_for_remove(new_identity_type, user_assigned)
+
+    target_identity = models_20220301preview.ManagedIdentityProperties()
+    target_identity.type=new_identity_type
+    target_identity.user_assigned_identities = user_identity_payload
+
+    app_resource = models_20220301preview.AppResource()
+    app_resource.identity = target_identity
+
+    poller = client.apps.begin_update(resource_group, service, name, app_resource)
+    wait_till_end(cmd, poller)
+    return client.apps.get(resource_group, service, name)
+
+
+def app_identity_show(cmd, client, resource_group, service, name):
+    app = client.apps.get(resource_group, service, name)
+    return app.identity
+
+
+def _is_legacy_identity_assign(system_assigned, user_assigned):
+    return not system_assigned and not user_assigned
+
+
+def _legacy_app_identity_assign(cmd, client, resource_group, service, name):
+    """
+    Enable system-assigned managed identity on app.
+    """
+    logger.warning("Start to enable system-assigned managed identity.")
+
+    app = client.apps.get(resource_group, service, name)
+    new_identity_type = models_20220301preview.ManagedIdentityType.SYSTEM_ASSIGNED
+    if app.identity and app.identity.type in (models_20220301preview.ManagedIdentityType.USER_ASSIGNED,
+                                              models_20220301preview.ManagedIdentityType.SYSTEM_ASSIGNED_USER_ASSIGNED):
+        new_identity_type = models_20220301preview.ManagedIdentityType.SYSTEM_ASSIGNED_USER_ASSIGNED
+    target_identity = models_20220301preview.ManagedIdentityProperties(type=new_identity_type)
+    app_resource = models_20220301preview.AppResource(identity=target_identity)
+
+    wait_till_end(cmd, client.apps.begin_update(resource_group, service, name, app_resource))
+
+
+def _new_app_identity_assign(cmd, client, resource_group, service, name, system_assigned, user_assigned):
+    app = client.apps.get(resource_group, service, name)
+    new_identity_type = _get_new_identity_type_for_assign(app, system_assigned, user_assigned)
+    user_identity_payload = _get_user_identity_payload_for_assign(new_identity_type, user_assigned)
+
+    identity_payload = models_20220301preview.ManagedIdentityProperties()
+    identity_payload.type = new_identity_type
+    identity_payload.user_assigned_identities = user_identity_payload
+
+    app_resource = models_20220301preview.AppResource(identity=identity_payload)
+
+    logger.warning("Start to assign managed identities to app.")
+    wait_till_end(cmd, client.apps.begin_update(resource_group, service, name, app_resource))
+
+
+def _get_new_identity_type_for_assign(app, system_assigned, user_assigned):
+    new_identity_type = None
+
+    if app.identity and app.identity.type:
+        new_identity_type = app.identity.type
+    else:
+        new_identity_type = models_20220301preview.ManagedIdentityType.NONE
+
+    if system_assigned:
+        if new_identity_type in (models_20220301preview.ManagedIdentityType.USER_ASSIGNED,
+                                 models_20220301preview.ManagedIdentityType.SYSTEM_ASSIGNED_USER_ASSIGNED):
+            new_identity_type = models_20220301preview.ManagedIdentityType.SYSTEM_ASSIGNED_USER_ASSIGNED
+        else:
+            new_identity_type = models_20220301preview.ManagedIdentityType.SYSTEM_ASSIGNED
+
+    if user_assigned:
+        if new_identity_type in (models_20220301preview.ManagedIdentityType.SYSTEM_ASSIGNED,
+                                 models_20220301preview.ManagedIdentityType.SYSTEM_ASSIGNED_USER_ASSIGNED):
+            new_identity_type = models_20220301preview.ManagedIdentityType.SYSTEM_ASSIGNED_USER_ASSIGNED
+        else:
+            new_identity_type = models_20220301preview.ManagedIdentityType.USER_ASSIGNED
+
+    if not new_identity_type or new_identity_type == models_20220301preview.ManagedIdentityType.NONE:
+        raise CLIError("Internal error: invalid new identity type:{}.".format(new_identity_type))
+
+    return new_identity_type
+
+
+def _get_user_identity_payload_for_assign(new_identity_type, new_user_identity_rid_list):
+    """
+    :param new_user_identity_rid_list: 1. None object.
+                                       2. A non-empty list of user-assigned managed identity resource ID.
+    :return 1. None object.
+            2. A dict from user-assigned managed identity to an empty object.
+    """
+    uid_payload = {}
+    if new_identity_type == models_20220301preview.ManagedIdentityType.SYSTEM_ASSIGNED:
+        pass
+    elif new_identity_type in (models_20220301preview.ManagedIdentityType.USER_ASSIGNED,
+                               models_20220301preview.ManagedIdentityType.SYSTEM_ASSIGNED_USER_ASSIGNED):
+        if new_user_identity_rid_list:
+            for rid in new_user_identity_rid_list:
+                uid_payload[rid]= models_20220301preview.UserAssignedManagedIdentity()
+
+    if len(uid_payload) == 0:
+        uid_payload = None
+
+    return uid_payload
+
+
+def _create_role_assignment(resource_group, service, name, role, scope):
+    app = client.apps.get(resource_group, service, name)
+
+    if not app.identity or not app.identity.principal_id:
+        raise CLIError("Failed to create role assignment without object ID(principal ID) of system-assigned managed identity.")
+
+    identity_role_id = _arm.resolve_role_id(cmd.cli_ctx, role, scope)
+    assignments_client = get_mgmt_service_client(cmd.cli_ctx, ResourceType.MGMT_AUTHORIZATION).role_assignments
+    RoleAssignmentCreateParameters = get_sdk(cmd.cli_ctx, ResourceType.MGMT_AUTHORIZATION,
+                                             'RoleAssignmentCreateParameters', mod='models',
+                                             operation_group='role_assignments')
+    parameters = RoleAssignmentCreateParameters(role_definition_id=identity_role_id, principal_id=principal_id)
+    logger.warning("Creating an assignment with a role '%s' on the scope of '%s'", identity_role_id, scope)
+    retry_times = 36
+    assignment_name = _arm._gen_guid()
+    for i in range(0, retry_times):
+        try:
+            assignments_client.create(scope=scope, role_assignment_name=assignment_name,
+                                      parameters=parameters)
+            break
+        except HttpResponseError as ex:
+            if 'role assignment already exists' in ex.message:
+                logger.warning('Role assignment already exists')
+                break
+            elif i < retry_times and ' does not exist in the directory ' in ex.message:
+                sleep(APP_CREATE_OR_UPDATE_SLEEP_INTERVAL)
+                logger.warning('Retrying role assignment creation: %s/%s', i + 1,
+                               retry_times)
+                continue
+            else:
+                raise
+
+
+def _get_new_user_identities_for_remove(exist_user_identity_dict, user_identity_list_to_remove):
+    """
+    :param exist_user_identity_dict: A dict from user-assigned managed identity resource id to identity objecct.
+    :param user_identity_list_to_remove: None, an empty list or a list of string of user-assigned managed identity resource id to remove.
+    :return A list of string of user-assigned managed identity resource ID.
+    """
+    if not exist_user_identity_dict:
+        return []
+
+    # None
+    if user_identity_list_to_remove is None:
+        return list(exist_user_identity_dict.keys())
+
+    # Empty list means remove all user-assigned managed identities
+    if len(user_identity_list_to_remove) == 0:
+        return []
+
+    # Non-empty list
+    new_identities = []
+    for id in exist_user_identity_dict.keys():
+        if not id.lower() in user_identity_list_to_remove:
+            new_identities.append(id)
+
+    return new_identities
+
+
+def _get_new_identity_type_for_remove(exist_identity_type, is_remove_system_identity, new_user_identities):
+    new_identity_type = exist_identity_type
+
+    exist_identity_type_str = exist_identity_type.lower()
+
+    if exist_identity_type_str == models_20220301preview.ManagedIdentityType.NONE.lower():
+        new_identity_type = models_20220301preview.ManagedIdentityType.NONE
+    elif exist_identity_type_str == models_20220301preview.ManagedIdentityType.SYSTEM_ASSIGNED.lower():
+        if is_remove_system_identity:
+            new_identity_type = models_20220301preview.ManagedIdentityType.NONE
+        else:
+            new_identity_type = models_20220301preview.ManagedIdentityType.SYSTEM_ASSIGNED
+    elif exist_identity_type_str == models_20220301preview.ManagedIdentityType.USER_ASSIGNED.lower():
+        if not new_user_identities:
+            new_identity_type = models_20220301preview.ManagedIdentityType.NONE
+        else:
+            new_identity_type = models_20220301preview.ManagedIdentityType.USER_ASSIGNED
+    elif exist_identity_type_str == models_20220301preview.ManagedIdentityType.SYSTEM_ASSIGNED_USER_ASSIGNED.lower():
+        if is_remove_system_identity and not new_user_identities:
+            new_identity_type = models_20220301preview.ManagedIdentityType.NONE
+        elif not is_remove_system_identity and not new_user_identities:
+            new_identity_type = models_20220301preview.ManagedIdentityType.SYSTEM_ASSIGNED
+        elif is_remove_system_identity and new_user_identities:
+            new_identity_type = models_20220301preview.ManagedIdentityType.USER_ASSIGNED
+        else:
+            new_identity_type = models_20220301preview.ManagedIdentityType.SYSTEM_ASSIGNED_USER_ASSIGNED
+    else:
+        raise CLIError("Invalid identity type: {}.".format(exist_identity_type_str))
+
+    return new_identity_type
+
+
+def _get_user_identity_payload_for_remove(new_identity_type, user_identity_list_to_remove):
+    """
+    :param new_identity_type: ManagedIdentityType
+    :param user_identity_list_to_remove: None, an empty list or a list of string of user-assigned managed identity resource id to remove.
+    :return None object or a non-empty dict from user-assigned managed identity resource id to None object
+    """
+    user_identity_payload = {}
+    if new_identity_type in (models_20220301preview.ManagedIdentityType.USER_ASSIGNED,
+                               models_20220301preview.ManagedIdentityType.SYSTEM_ASSIGNED_USER_ASSIGNED):
+        # empty list means remove all user-assigned managed identites
+        if user_identity_list_to_remove is not None and len(user_identity_list_to_remove) == 0:
+            raise CLIError("Target identity type should not be {}, when remove all user-assigned managed identities.".format(new_identity_type))
+        # non-empty list
+        elif user_identity_list_to_remove:
+            for id in user_identity_list_to_remove:
+                user_identity_payload[id] = None
+
+    if not user_identity_payload:
+        user_identity_payload = None
+
+    return user_identity_payload
+
